Moved common tor disposable logic to base class
