+++ WalletWasabi/WebClients/ChaumianCoinJoin/AliceClient.cs
@@ -13,21 +13,19 @@ using WalletWasabi.Backend.Models.Responses;
 using WalletWasabi.Logging;
 using WalletWasabi.Models.ChaumianCoinJoin;
 using WalletWasabi.TorSocks5;
+using WalletWasabi.Bases;
 
 namespace WalletWasabi.WebClients.ChaumianCoinJoin
 {
-	public class AliceClient : IDisposable
+	public class AliceClient : TorDisposableBase
 	{
-		public TorHttpClient TorClient { get; }
 
 		public long RoundId { get; private set; }
 		public Guid UniqueId { get; private set; }
 		public byte[] BlindedOutputSignature { get; private set; }
-
-		/// <param name="torSocks5EndPoint">if null, then localhost:9050</param>
-		private AliceClient(Uri baseUri, IPEndPoint torSocks5EndPoint = null)
+		/// <inheritdoc/>
+		private AliceClient(Uri baseUri, IPEndPoint torSocks5EndPoint = null) : base(baseUri, torSocks5EndPoint)
 		{
-			TorClient = new TorHttpClient(baseUri, torSocks5EndPoint, isolateStream: true);
 		}
 
 		public static async Task<AliceClient> CreateNewAsync(InputsRequest request, Uri baseUri, IPEndPoint torSocks5EndPoint = null)
@@ -147,37 +145,5 @@ namespace WalletWasabi.WebClients.ChaumianCoinJoin
 				Logger.LogInfo<AliceClient>($"Round ({RoundId}), Alice ({UniqueId}): Posted {signatures.Count} signatures.");
 			}
 		}
-
-		#region IDisposable Support
-
-		private volatile bool _disposedValue = false; // To detect redundant calls
-
-		protected virtual void Dispose(bool disposing)
-		{
-			if (!_disposedValue)
-			{
-				if (disposing)
-				{
-					TorClient?.Dispose();
-				}
-
-				_disposedValue = true;
-			}
-		}
-
-		// ~AliceClient() {
-		//   // Do not change this code. Put cleanup code in Dispose(bool disposing) above.
-		//   Dispose(false);
-		// }
-
-		// This code added to correctly implement the disposable pattern.
-		public void Dispose()
-		{
-			// Do not change this code. Put cleanup code in Dispose(bool disposing) above.
-			Dispose(true);
-			// GC.SuppressFinalize(this);
-		}
-
-		#endregion IDisposable Support
 	}
 }
